Added rendering for the water surface
