+++ Graphics Engine/Graphics Engine.vcxproj
@@ -176,6 +176,9 @@
     <ClCompile Include="src\datamodel\terrain\Terrain.cpp" />
     <ClCompile Include="src\core\ThreadPool.cpp" />
     <ClCompile Include="src\physics\PhysicsTerrain.cpp" />
+    <ClCompile Include="src\rendering\terrain\BufferPool.cpp" />
+    <ClCompile Include="src\rendering\terrain\WaterSurface.cpp" />
+    <ClCompile Include="src\rendering\terrain\VisualTerrainCallback.cpp" />
     <ClCompile Include="src\_DEP_\TreeGenerator.cpp" />
     <ClCompile Include="src\input\components\MovementHandler.cpp" />
     <ClCompile Include="src\input\InputFrame.cpp" />
@@ -237,7 +240,7 @@
     <ClCompile Include="src\physics\PhysicsSystem.cpp" />
     <ClCompile Include="src\rendering\VisualDebug.cpp" />
     <ClCompile Include="src\rendering\VisualSystem.cpp" />
-    <ClCompile Include="src\rendering\VisualTerrain.cpp" />
+    <ClCompile Include="src\rendering\terrain\VisualTerrain.cpp" />
     <ClCompile Include="src\utility\FileReader.cpp" />
     <ClCompile Include="src\utility\Stopwatch.cpp" />
     <ClCompile Include="src\rendering\core\VertexStreamIDs.h" />
@@ -253,6 +256,9 @@
     <ClInclude Include="src\core\ThreadPool.h" />
     <ClInclude Include="src\datamodel\terrain\TerrainCallback.h" />
     <ClInclude Include="src\physics\PhysicsTerrain.h" />
+    <ClInclude Include="src\rendering\terrain\BufferPool.h" />
+    <ClInclude Include="src\rendering\terrain\WaterSurface.h" />
+    <ClInclude Include="src\rendering\terrain\VisualTerrainCallback.h" />
     <ClInclude Include="src\_DEP_\ResourceManager.h" />
     <ClInclude Include="src\datamodel\SceneGraph.h" />
     <ClInclude Include="src\datamodel\terrain\MarchingCube.h" />
@@ -323,7 +329,7 @@
     <ClInclude Include="src\rendering\Direct3D11.h" />
     <ClInclude Include="src\rendering\VisualDebug.h" />
     <ClInclude Include="src\rendering\VisualSystem.h" />
-    <ClInclude Include="src\rendering\VisualTerrain.h" />
+    <ClInclude Include="src\rendering\terrain\VisualTerrain.h" />
     <ClInclude Include="src\utility\FileReader.h" />
     <ClInclude Include="src\utility\Stopwatch.h" />
   </ItemGroup>
