+++ Graphics Engine/imgui.ini
@@ -1,6 +1,6 @@
 [Window][Debug##Default]
-Pos=223,0
-Size=804,146
+Pos=218,3
+Size=804,218
 Collapsed=1
 
 [Window][Dear ImGui Demo]
