+++ Graphics Engine/src/rendering/resources/ResourceManager.cpp
@@ -30,10 +30,17 @@ ResourceManager::~ResourceManager() = default;
 
 // Initialize:
 // Loads assets into the asset manager.
+static ID3D11DepthStencilState* LoadDSTestAndWrite(ID3D11Device* device);
+static ID3D11DepthStencilState* LoadDSTestNoWrite(ID3D11Device* device);
+
 void ResourceManager::initializeResources() {
     // Create my samplers
     shadowmap_sampler = LoadShadowMapSampler();
     mesh_sampler = LoadMeshTextureSampler();
+
+    // Create my depth stencil states
+    ds_test_and_write = LoadDSTestAndWrite(device);
+    ds_test_no_write = LoadDSTestNoWrite(device);
 
     // Stores an atlas of material colors to avoid the need for rebinds later
     AtlasBuilder atlas_builder = AtlasBuilder(4096, 4096);
@@ -102,6 +109,13 @@ ID3D11SamplerState* ResourceManager::getShadowMapSampler() {
 }
 
 ID3D11SamplerState* ResourceManager::getMeshSampler() { return mesh_sampler; }
+
+ID3D11DepthStencilState* ResourceManager::DSState_TestNoWrite() {
+    return ds_test_no_write;
+}
+ID3D11DepthStencilState* ResourceManager::DSState_TestAndWrite() {
+    return ds_test_and_write;
+}
 
 // LoadAssetFromGLTF:
 // Uses the GLTFFile interface to load an asset from a GLTF file
@@ -197,5 +211,42 @@ ID3D11SamplerState* ResourceManager::LoadMeshTextureSampler() {
 
     return sampler;
 }
+
+ID3D11DepthStencilState* LoadDSTestAndWrite(ID3D11Device* device) {
+    D3D11_DEPTH_STENCIL_DESC desc = {};
+    // Enable depth testing
+    desc.DepthEnable = TRUE;
+    // Standard depth test
+    desc.DepthFunc = D3D11_COMPARISON_LESS;
+    // Enable depth writing
+    desc.DepthWriteMask = D3D11_DEPTH_WRITE_MASK_ALL;
+    // No stencil testing
+    desc.StencilEnable = FALSE;
+
+    ID3D11DepthStencilState* state = nullptr;
+    HRESULT result = device->CreateDepthStencilState(&desc, &state);
+    assert(SUCCEEDED(result));
+
+    return state;
+}
+
+ID3D11DepthStencilState* LoadDSTestNoWrite(ID3D11Device* device) {
+    D3D11_DEPTH_STENCIL_DESC desc = {};
+    // Enable depth testing
+    desc.DepthEnable = TRUE;
+    // Standard depth test
+    desc.DepthFunc = D3D11_COMPARISON_LESS;
+    // Disable depth writing
+    desc.DepthWriteMask = D3D11_DEPTH_WRITE_MASK_ZERO;
+    // No stencil testing
+    desc.StencilEnable = FALSE;
+
+    ID3D11DepthStencilState* state = nullptr;
+    HRESULT result = device->CreateDepthStencilState(&desc, &state);
+    assert(SUCCEEDED(result));
+
+    return state;
+}
+
 } // namespace Graphics
 } // namespace Engine
