+++ Graphics Engine/src/rendering/shaders/ShaderManager.cpp
@@ -132,6 +132,11 @@ void ShaderManager::initializeShaders() {
         {"LightFrustum", "V_LightFrustum.hlsl", "vs_main", false});
     createPixelShader({"LightFrustum", "P_LightFrustum.hlsl", "ps_main"});
 
+    input_layout_arr = {POSITION};
+    createVertexShader(
+        {"WaterSurface", "V_WaterSurface.hlsl", "vs_main", false});
+    createPixelShader({"WaterSurface", "P_WaterSurface.hlsl", "ps_main"});
+
     // --- Post Processing Effects ---
     // Generic vertex shader for post process effects
     input_layout_arr = {SV_POSITION};
